requirements.txt
streamlit==1.42.2
pandas==2.2.3
pydeck==0.9.1
numpy==1.26.4
psycopg2-binary==2.9.10
httpx==0.28.0
datetime==5.5
sqlalchemy==2.0.38
altair==5.5.0
urllib3==2.2.2
requests==2.32.3
openmeteo_requests==1.3.0
requests_cache==1.2.1
retry_requests==2.0.0
